backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
cohere==4.4.3
qdrant-client==1.8.0
python-dotenv==1.0.0
pytest==7.4.3